requirements.txt
requests>=2.31.0
graphrag>=0.1.1
numpy>=1.24.2
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.2.4
pyyaml>=6.0
ipykernel==6.26.0
tiktoken
tabulate
ipywidgets
tqdm